backend/requirements.txt
# Django and API Framework
Django==5.0.2
djangorestframework==3.15.0
django-cors-headers==4.3.1

# Database support
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# File handling
python-multipart>=0.0.6
PyPDF2>=3.0.0,<4.0
pdfplumber>=0.7.0,<1.0

# Environment and configuration
python-dotenv==1.0.1

# AI processing
openai>=1.0.0
portkey-ai==1.14.1
pydantic==2.6.1

# Image processing
Pillow==10.2.0
pdf2image>=1.16.0,<2.0
pytesseract>=0.3.10

# Cloud storage
boto3==1.34.50

# Browser automation
playwright==1.40.0
asyncio-throttle>=1.0.2

# Additional dependencies
requests>=2.28.0
aiohttp>=3.8.0